+++ src/main/java/gift/auth/OAuthToken.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.auth;
 
 public record OAuthToken (
         String access_token,
