+++ src/main/java/gift/component/kakao/KakaoProperties.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.component.kakao;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
