+++ src/main/java/gift/dto/KakaoUserInfoResponseDto.java
@@ -1,10 +1,10 @@
-package gift;
+package gift.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserInfoResponse(
+public record KakaoUserInfoResponseDto(
         Long id
 ) {
 }
