+++ src/main/java/gift/service/MemberService.java
@@ -1,5 +1,6 @@
 package gift.service;
 
+import gift.auth.JwtHelper;
 import gift.repository.MemberRepository;
 import gift.vo.Member;
 import org.springframework.stereotype.Service;
